loggedInView => appView
